Restructured and refactored Profile emails to now be stored in one oneToMany relationship. Primary emails are now known by a boolean flag on the the email which is checked for when requesting a primary email from a profile. This is not the most elegant way of doing things as it is possible to invalid states within the database.
For instance it is now possible if state is not maintained to have a profile that could have multiple possible primary emails. Right now this is dealt with by a first seen first valid check however we could add a constraint to the database itself to only have one primary for a collection of emails with the same profile id. I am unsure how to do this right now but may be worth being a new task in the future.

An important change to this design is now Profile.getEmails returns all emails not just additional emails. If this is incorrect semantics it would only require filtering the emails that are non primary.

Changes:
+ Modified Profile to use just one email list
+ Modfied Email to have primary column
+ Modified ProfileRepository to use custom sql queries instead of auto create to accomplish tasks done before.
+ Refactored Profile email getters and setters naming introducing smalls changes throughout the codebase.