Implement the loggedinuserdata function that loads in the logged in user on edit profile.
This allows the navbar to go to the logged in users profile when clicking 'profile' even when on another users page. Currently working on a solution to not allow users to manually enter url to go to another users edit profile page. having difficulty with routing #implement #pair[mlo53, ewm27] #story[256]