Modified editActivity endpoint to properly deal with the change of visibility
I created a query method in activityRoleRepo deleteAllAccessinActivity and a query method in subscriptionHistoryRepo called unsubscribeAllButCreator. These methods help with dealing with the logic when the visibility is changed in the editActivity endpoint. It behaves a little differently to the editaVisbility endpoint as it doesnt have an accessors list. #implement #story[340]