Changed if statmenets to properly compare strings which fixed a test.
Also added a check for accessors in the request if changing to public or private. The accessors should be empty so if it isnt sends a bad request back. fixed some test that needed to set subscriber method when subscribing. Deleted old test that doesn't work for API spec in editActivityTest. #implement #fix #test #story[340]